feat: use string enums over numeric
Oddly, Twirp RPC responses use string enums, and Twirp RPC requests have to use
numeric. Since we're having to bridge a gap on numeric vs string anyway,
i've switched the default to string.

With this comes easier readability, one of the features of string
enums as described by TypeScript docs. And, readability and usability
is the main goal of this lib.